Add Changes 13.01.22
Remove VS Code Folder
Reposition VS Folder
Restore Input Validation (Now ONLY Number acceptable)
