--- ASCII_TIMER/ASCII_TIMER_VS/ASCII_TIMER_VS/other_functions.cpp
@@ -1,101 +1,0 @@
-/**
- *
- * Solution to course project # 9
- * Introduction to programming course
- * Faculty of Mathematics and Informatics of Sofia University
- * Winter semester 2021/2022
- *
- * @author Filip Filchev
- * @idnumber 0MI0600041
- * @compiler VC
- *
- * File with all Functions that don't fit anywhere else
- *
- */
-#include "other_functions.h"
-#include <iostream>
-using namespace std;
-void screenClear() {
-    // Clear the Console
-#ifdef _WIN32
-    system("cls");
-#elif __linux__
-    system("clear");
-#else
-    for (int i = 0; i <= 200; ++i) {
-        cout << endl;
-    }
-#endif
-}
-int tensOfNumberCount(const int n) {
-    // Return the Count of Tens is the number
-    int count = 1;
-    for (int i = n; i > 0; i /= 10) {
-        count *= 10;
-    }
-    count /= 10;
-    return count;
-}
-int *transformHoursMinutesSeconds(int n) {
-    // Transforming the Input
-    int *timeArr = new int[3];
-    timeArr[0] = 0; // Hours
-    timeArr[1] = 0; // Minutes
-    timeArr[2] = 0; // Seconds
-    if (n < 60) {
-        timeArr[2] = n;
-    } else {
-        for (int i = n; i > 0; i -= 60) {
-            if (n >= 60) {
-                ++timeArr[1];
-            }
-            n -= 60;
-            if (timeArr[1] == 60) {
-                timeArr[1] = 0;
-                ++timeArr[0];
-            }
-        }
-        if (n != 0) {
-            timeArr[2] = 60 + n;
-        }
-    }
-    return timeArr;
-}
-bool isInputCorrect(string input) {
-    // Validate the Input
-    for (int i = 0; input[i] != '\0'; ++i) {
-        if (input[i] < '0' || input[i] > '9') {
-            return false;
-        }
-        if (input[i] == ' ' || input[i] == '\0' || input[i] == '\t') {
-            if (input[i + 1] == ' ' || input[i + 1] == '\0' || input[i + 1] == '\t') {
-                return false;
-            }
-
-            int countOfDigits = 0;
-            for (int j = i + 1; input[j] != ' '; ++j) {
-                if (input[j] == '\0') {
-                    break;
-                }
-                if (input[j] >= '0' && input[j] <= '9') {
-                    ++countOfDigits;
-                }
-            }
-            if (countOfDigits != 3) { // Valid are only Inputs like 1 234 567 890 or 1234567890
-                return false;
-            }
-        }
-    }
-    return true;
-}
-int getNumber(string input) {
-    // Transform string to integer
-    int number = 0;
-    for (int i = 0; input[i] != '\0'; ++i) {
-        if (input[i] == ' ') {
-            continue;
-        }
-        number = number * 10 + (input[i] - '0');
-    }
-    return number;
-}
