+++ ASCII_TIMER/ASCII_TIMER_VS/TestsInputs.txt
@@ -1,16 +1,17 @@
 Valid Inputs:
 
-1 000 000 000
-1 000 000
-100 000
-10 000
-1 000
+1000000000
+1000000
+100000
+10000
+1000
 100
 ===========================================================================
+
 Invalid Inputs::
 
 1. All Types of Characters other than Numbers
-2. Only Numbers like:
+2. Numbers like:
 | 1 2
 | 1 23
 | 12 3
@@ -21,10 +22,14 @@ Invalid Inputs::
 | 4567 890
 | 1 23 456
 | 12 3 456
-Aka. Numbers that start with a Digit and don't have exactly 3 digits after each Space
+| 1 234 567 890
+tl:dr If the Input has atleast one Character other than a Number
 ===========================================================================
+
 Tests:
 
+//These Inputs are Invalid, but it's easier to read
+=========================================================================== 
 3 600 000| -> 1000 Hours (Transform to 3 Digit Hours) Centering Test
 360 000--|--> 100 Hours (Transform to 2 Digit Hours) Centering Test
 36 000---|--> 10 Hours (Transform to single Digit Hours) Centering Test
