+++ ASCII_TIMER/ASCII_TIMER_VS/ascii_timer.cpp
