+++ ASCII_TIMER/ASCII_TIMER_VS/other_functions.h
