--- ASCII_TIMER/ASCII_Timer_CODE/ascii_timer.cpp
@@ -1,6 +1,0 @@
-#include "home_screen.h"
-
-int main() {
-	homeScreen();
-	return 0;
-}
