--- ASCII_TIMER/ASCII_Timer_CODE/end_screen.h
@@ -1,6 +1,0 @@
-#pragma once
-#ifndef END_SCREEN_H
-#define END_SCREEN_H
-void printEndScreen(bool invalidEntry);
-void endScreen(bool invalidEntry);
-#endif
