--- ASCII_TIMER/ASCII_Timer_CODE/home_screen.h
@@ -1,11 +1,0 @@
-#pragma once
-#ifndef HOME_SCREEN_H
-#define HOME_SCREEN_H
-void homeScreen();
-void printHomeScreenAndMenu();
-void printT(int line, int color);
-void printI(int color);
-void printM(int line, int color);
-void printE(int line, int color);
-void printR(int line, int color);
-#endif
