--- ASCII_TIMER/ASCII_Timer_CODE/other_functions.h
@@ -1,7 +1,0 @@
-#pragma once
-#ifndef OTHER_FUNCTIONS_H
-#define OTHER_FUNCTIONS_H
-void screenClear();
-int tensOfNumberCount(int n);
-int* transformHoursMinutesSeconds(int n);
-#endif
