--- ASCII_TIMER/ASCII_Timer_CODE/print_hours.h
@@ -1,5 +1,0 @@
-#pragma once
-#ifndef PRINT_HOURS_H
-#define PRINT_HOURS_H
-void printHours(int hours, int lineNumber, int tensOfNumberCount, bool isHoursLessThanTen, int color1, int color2);
-#endif
