--- ASCII_TIMER/ASCII_Timer_CODE/print_minutes.h
@@ -1,5 +1,0 @@
-#pragma once
-#ifndef PRINT_MINUTES_H
-#define PRINT_MINUTES_H
-void printMinutes(const int minutes, const int lineNumber, const int color1, const int color2);
-#endif
