--- ASCII_TIMER/ASCII_Timer_CODE/print_numbers.h
@@ -1,14 +1,0 @@
-#pragma once
-#ifndef PRINT_NUMBERS_H
-#define PRINT_NUMBERS_H
-void print0(int num, int lineNumber, bool isItTheSecondDigit);
-void print1(int num, int lineNumber, bool isItTheSecondDigit);
-void print2(int num, int lineNumber, bool isItTheSecondDigit);
-void print3(int num, int lineNumber, bool isItTheSecondDigit);
-void print4(int num, int lineNumber, bool isItTheSecondDigit);
-void print5(int num, int lineNumber, bool isItTheSecondDigit);
-void print6(int num, int lineNumber, bool isItTheSecondDigit);
-void print7(int num, int lineNumber, bool isItTheSecondDigit);
-void print8(int num, int lineNumber, bool isItTheSecondDigit);
-void print9(int num, int lineNumber, bool isItTheSecondDigit);
-#endif
