--- ASCII_TIMER/ASCII_Timer_CODE/print_seconds.h
@@ -1,5 +1,0 @@
-#pragma once
-#ifndef PRINT_SECONDS_H
-#define PRINT_SECONDS_H
-void printSeconds(const int seconds, const int lineNumber, int color1, int color2);
-#endif
